Project Structure Setup
This repo mimicks the layout of the admin panel where custom html, js & css can be applied to make it simpler for site admins to upload overrides.

Note that guide.html and guide-alt.html are default LibGuides html templates. guide-alt.html is not in use at the moment and can be used for testing in a live environment without disruption.