Development Workflow Setup, Bootstrap Nav Example
Hacked a development environment into LibGuides. Dev work happens entirely in guide-alt.html, and can be pushed to the rest of the site once we're ready.

Note that having bootstrap links & scripts in guide-alt.html breaks it's LibGuides editor page. They must be commented out to edit the content of the page with LibGuide tools.

Also copied existing css tweaks to head.html