Header Nav in Production, Additional Nav tweaks
Navigation bar is now in production. Previous nav bar has been hidden.

Also made more use of dropdowns in the final version.

On the LibGuides side, work has been done to populate the a-z databases and subject categories that the nav bar points to. Some of that work can be tracked on github. Most of it cannot.